requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
opencv-python==4.8.1.78
numpy==1.21.6
Pillow==10.0.0
pyttsx3==2.90
python-dotenv==1.0.0
Werkzeug==2.3.7
scipy==1.9.3
psycopg2-binary==2.9.9